+++ src/backend/SvgBackend.h
@@ -18,9 +18,9 @@ class SvgBackend : public Backend
     PdfSvgClosure* _closure = NULL;
     inline PdfSvgClosure* closure() { return _closure; }
 
-    SvgBackend(int width, int height);
+    SvgBackend(double width, double height);
     ~SvgBackend();
-    static Backend *construct(int width, int height);
+    static Backend *construct(double width, double height);
 
     static Nan::Persistent<v8::FunctionTemplate> constructor;
     static void Initialize(v8::Handle<v8::Object> target);
